requirements.txt
requests
discord.py 
python-dotenv
openai
youtube_dl
